# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Geekron CMS API
  description: |
    Geekron CMS 多租户内容管理系统 API
    
    ## 认证
    大部分接口需要 JWT Token 认证，在请求头中添加：
    ```
    Authorization: Bearer <your_token>
    ```
  version: 1.0.0
  contact:
    name: Geekron CMS Team
    email: support@geekron-cms.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: https://api.geekron-cms.com
    description: Production
  - url: https://staging-api.geekron-cms.com
    description: Staging
  - url: http://localhost:8787
    description: Local Development

tags:
  - name: Authentication
    description: 认证相关接口
  - name: Tenants
    description: 租户管理
  - name: Users
    description: 用户管理
  - name: Content
    description: 内容管理
  - name: Models
    description: 数据模型
  - name: Files
    description: 文件存储

paths:
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: 用户登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: 认证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/auth/register:
    post:
      tags:
        - Authentication
      summary: 用户注册
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/tenants:
    get:
      tags:
        - Tenants
      summary: 获取租户列表
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    post:
      tags:
        - Tenants
      summary: 创建租户
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/tenants/{id}:
    get:
      tags:
        - Tenants
      summary: 获取租户详情
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    
    put:
      tags:
        - Tenants
      summary: 更新租户
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTenantRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      tags:
        - Tenants
      summary: 删除租户
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: 删除成功
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/users:
    get:
      tags:
        - Users
      summary: 获取用户列表
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
    
    post:
      tags:
        - Users
      summary: 创建用户
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

  /api/v1/content:
    get:
      tags:
        - Content
      summary: 获取内容列表
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: model
          in: query
          schema:
            type: string
          description: 按模型筛选
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, archived]
          description: 按状态筛选
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentListResponse'
    
    post:
      tags:
        - Content
      summary: 创建内容
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContentRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentResponse'

  /api/v1/files/upload:
    post:
      tags:
        - Files
      summary: 上传文件
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: 上传成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
      description: 页码
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
      description: 每页数量
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: 资源 ID

  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
    
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            token:
              type: string
            expiresIn:
              type: integer
            user:
              $ref: '#/components/schemas/User'
    
    Tenant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        subdomain:
          type: string
        status:
          type: string
          enum: [active, suspended, deleted]
        plan:
          type: string
          enum: [free, pro, enterprise]
        settings:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    
    CreateTenantRequest:
      type: object
      required:
        - name
        - subdomain
      properties:
        name:
          type: string
        subdomain:
          type: string
          pattern: '^[a-z0-9-]+$'
        plan:
          type: string
          enum: [free, pro, enterprise]
          default: free
        settings:
          type: object
    
    UpdateTenantRequest:
      type: object
      properties:
        name:
          type: string
        plan:
          type: string
        settings:
          type: object
    
    TenantResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Tenant'
    
    TenantListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Tenant'
        pagination:
          $ref: '#/components/schemas/Pagination'
    
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [admin, member, viewer]
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
    
    CreateUserRequest:
      type: object
      required:
        - email
        - name
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [admin, member, viewer]
          default: member
    
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/User'
    
    UserListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          $ref: '#/components/schemas/Pagination'
    
    Content:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
        modelId:
          type: string
        data:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    
    CreateContentRequest:
      type: object
      required:
        - title
        - modelId
      properties:
        title:
          type: string
        slug:
          type: string
        modelId:
          type: string
        data:
          type: object
        status:
          type: string
          enum: [draft, published]
          default: draft
    
    ContentResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Content'
    
    ContentListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Content'
        pagination:
          $ref: '#/components/schemas/Pagination'
    
    File:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        url:
          type: string
        size:
          type: integer
        mimeType:
          type: string
        createdAt:
          type: string
          format: date-time
    
    FileResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/File'
    
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    
    Error:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object

  responses:
    Unauthorized:
      description: 未授权
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'